Add files via upload
The Job Search Layout is an interactive platform that allows users to search for jobs based on their experience level. Featuring hover effects and a dropdown list selection, this project offers an intuitive and user-friendly interface for job seekers.

Built with ReactJS and Tailwind CSS, this application combines modern web technologies to deliver a seamless experience, empowering users to find opportunities that match their expertise.